rf: add sepparate files for each command
