+++ src/commands/index.js
@@ -1,0 +1,7 @@
+import { add } from "./add.js";
+import { cat } from "./cat.js";
+import { cd } from "./cd.js";
+import { ls } from "./ls.js";
+import { up } from "./up.js";
+
+export const actions = { ls, up, add, cat, cd };
